+++ src/main/java/cart/auth/AuthInfo.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.auth;
 
 public class AuthInfo {
